New upstream version 7.1.8-dfsg

+++ src/VBox/Runtime/r3/posix/timelocal-posix.cpp
@@ -4,7 +4,7 @@
  */
 
 /*
- * Copyright (C) 2006-2023 Oracle and/or its affiliates.
+ * Copyright (C) 2006-2024 Oracle and/or its affiliates.
  *
  * This file is part of VirtualBox base platform packages, as
  * available from https://www.virtualbox.org.
@@ -208,7 +208,7 @@ RTDECL(PRTTIME) RTTimeLocalExplode(PRTTIME pTime, PCRTTIMESPEC pTimeSpec)
     if (pTime)
     {
         pTime->fFlags = (pTime->fFlags & ~RTTIME_FLAGS_TYPE_MASK) | RTTIME_FLAGS_TYPE_LOCAL;
-        pTime->offUTC = cNsUtcOffset / RT_NS_1MIN;
+        pTime->offUTC = cNsUtcOffset / (int64_t)RT_NS_1MIN;
     }
     return pTime;
 }
